Try to implement hash operations

+++ trovefile.h
@@ -1,1 +1,1 @@
-extern unsigned long hashFunction(char *str);
+extern int hashFunction(char *str);
